+++ src/par.pxd
@@ -4,4 +4,4 @@ cdef extern from "par.h":
     cdef struct par3:
         size_t l[3]
         dbl b[3]
-    int par3_size(const par3 *par)
+    size_t par3_size(const par3 *par)
